+++ nilearn/datasets/__init__.py
@@ -21,7 +21,8 @@ from .atlas import (fetch_atlas_craddock_2012, fetch_atlas_destrieux_2009,
                     fetch_atlas_basc_multiscale_2015,
                     fetch_coords_dosenbach_2010,
                     fetch_atlas_allen_2011,
-                    fetch_atlas_surf_destrieux)
+                    fetch_atlas_surf_destrieux,
+                    fetch_atlas_talairach)
 
 from .utils import get_data_dirs
 from .neurovault import fetch_neurovault, fetch_neurovault_ids
@@ -43,5 +44,5 @@ __all__ = ['MNI152_FILE_PATH', 'fetch_icbm152_2009', 'load_mni152_template',
            'fetch_atlas_basc_multiscale_2015', 'fetch_coords_dosenbach_2010',
            'fetch_neurovault', 'fetch_neurovault_ids',
            'load_mni152_brain_mask', 'fetch_icbm152_brain_gm_mask',
-           'fetch_atlas_surf_destrieux',
+           'fetch_atlas_surf_destrieux', 'fetch_atlas_talairach',
            'get_data_dirs']